fixed Guardian Angel, Jump Boost WIP, Gas Disaster..
Some rendering fixes are in progress as well. I want this game to be fairly mod-friendly out of the box, and I'm working on putting fixes in place as a setup for texture packs later on.
Known issue: THE BACKGROUNDS HAVE BEEN A PIXEL TOO WIDE THIS WHOLE TIME, AND NOBODY WAS GONNA TELL ME!??!?! >:(